+++ android-app/app/src/main/res/layout/activity_main.xml
@@ -10,31 +10,21 @@
         android:id="@+id/frame_layout"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_below="@+id/popupbutton"
         android:layout_alignParentStart="true"
-        android:layout_marginTop="207dp" />
         android:layout_above="@+id/navigation"
         android:animateLayoutChanges="true">
     </FrameLayout>
-
     <android.support.design.widget.BottomNavigationView
         android:id="@+id/navigation"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_alignParentBottom="true"
-        android:layout_centerHorizontal="true" />
-    <Button
-        android:id="@+id/popupbutton"
-        android:layout_width="50dp"
-        android:layout_height="50dp"
-        android:background="@color/red_light"
-        android:layout_alignTop="@+id/button2"
-        android:layout_toEndOf="@+id/button2"
-        android:layout_marginStart="42dp" />
-        android:background="@color/colorPrimary"
-        android:theme="@style/AppTheme"
+        android:layout_centerHorizontal="true"
         app:itemIconTint="@color/white_smoke"
         app:itemTextColor="@color/white_smoke"
-        app:menu="@menu/bottom_navigation_items"/>
+        app:menu="@menu/bottom_navigation_items"
+        android:background="@color/colorPrimary"
+        android:theme="@style/AppTheme" />
+
 </RelativeLayout>
